Update folder structure, separate dto and entity
