--- src/main/java/repository/ToDoRepository.java
@@ -1,7 +1,0 @@
-package repository;
-
-import org.springframework.data.repository.CrudRepository;
-import web.Task;
-
-public interface ToDoRepository extends CrudRepository<Task, Integer> {
-}
